Ingrion_Web/backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
django-cors-headers==4.4.0
django-ratelimit==4.1.0
cryptography==42.0.8
Pillow==10.4.0
python-dotenv==1.0.1
gunicorn==22.0.0
whitenoise==6.7.0
